Refactor LOADROM out of dos_programs
